Reset logging console on log level change
In case the logging level is set to a more permissive/restrictive
value that the default one (`Level.INFO`), the console used to
only show the messages of that level only after the option
application.

After this change, the console will be updated with all the "missed"
log messages, in order to aid log reading and/or debugging.

The `out` channel will be left as-is, though, because there is no
way to "rewind" all the messages from a certain point onwards.

Signed-off-by: TheSilkMiner <sir.karati@outlook.com>